RF: Moved parse_url_opts under support.network
